Barf early when DOCKER_HOST is set on Linux.
As it stands now, if sphere-stack is running on Linux it is assumed that docker
is installed locally and that DOCKER_HOST is unnecessary.

Indeed, for the moment on Linux, we require that DOCKER_HOST is empty and that docker
is installed locally. If not, we barf early to explain the issue and solicit
requirements/votes for a better fix.

As a workaround, try:

   export DOCKER_HOST=
   ./sphere-stack ...


Signed-off-by: Jon Seymour <jon@wildducktheories.com>